Kernel: Remove Processor::is_kernel_mode
This function was only used to verify that we are running in kernel
mode. But it is pretty much impossible that we will ever end up in
kernel code and actually are able to execute it in user mode. A lot of
stuff must go completely wrong to end up in such a situation.

Getting the current privilege level is also impossible on RISC-V by
design.