LibWeb: Add a UIEvents/KeyCode.h forwarding header
It only forwards to Kernel/API/KeyCode.h for now. The main motivation
is being able to cherry-pick the ladybird change that put
`Web::UIEvents::` in front of all key event symbols, which in turn
will cause fewer conflicts when cherry-picking other things.

(It's arguably also a bit weird to consume kernel headers from
user space, but we currently do that pretty heavily.)