Kernel/x86: Don't attempt to parse the MultiProcessor table
We don't use its contents for anything. We only log the
PCI interrupt redirections defined by it.

The MultiProcessor configuration table is obsoleted by ACPI and not
present on most modern systems.

We currently tried to find it in BIOS regions even if we booted via EFI.
But those regions might not be present on (U)EFI systems, resulting in
crashes.