Kernel/riscv64: Don't VERIFY that we are running on hart 0
The PLIC driver now supports the boot hart ID not being 0.

We don't use the x86-64 APIC IDs as our CPU IDs, and I don't think we
should use the hart IDs on RISC-V for that purpose either, since they
aren't guaranteed to be contiguous.

The SiFive U74 (used in the JH7110) for example uses hart ID 0 for the
S7 monitor core, which doesn't implement S-mode, so we can't use it.